// frontend/src/generated/models/riskWorkspaceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CFO Command Center API
 * OpenAPI spec version: 0.3.0
 */
import type { RiskWorkspaceResponseAppetiteRadarItem } from './riskWorkspaceResponseAppetiteRadarItem';
import type { RiskWorkspaceResponseAssurance } from './riskWorkspaceResponseAssurance';
import type { RiskWorkspaceResponseCategoriesItem } from './riskWorkspaceResponseCategoriesItem';
import type { WorkspaceContextResponse } from './workspaceContextResponse';
import type { RiskWorkspaceResponseControlsItem } from './riskWorkspaceResponseControlsItem';
import type { RiskWorkspaceResponseCorrelation } from './riskWorkspaceResponseCorrelation';
import type { RiskWorkspaceResponseLineage } from './riskWorkspaceResponseLineage';
import type { RiskWorkspaceResponsePercentileCurveItem } from './riskWorkspaceResponsePercentileCurveItem';
import type { RiskWorkspaceResponsePortfolio } from './riskWorkspaceResponsePortfolio';
import type { RiskWorkspaceResponseRegimes } from './riskWorkspaceResponseRegimes';
import type { RiskWorkspaceResponseRisksItem } from './riskWorkspaceResponseRisksItem';
import type { RiskWorkspaceResponseScenario } from './riskWorkspaceResponseScenario';
import type { RiskWorkspaceResponseTail } from './riskWorkspaceResponseTail';

export interface RiskWorkspaceResponse {
  appetite_radar?: RiskWorkspaceResponseAppetiteRadarItem[];
  as_of: string;
  assurance?: RiskWorkspaceResponseAssurance;
  categories?: RiskWorkspaceResponseCategoriesItem[];
  context: WorkspaceContextResponse;
  controls?: RiskWorkspaceResponseControlsItem[];
  correlation?: RiskWorkspaceResponseCorrelation;
  lineage?: RiskWorkspaceResponseLineage;
  percentile_curve?: RiskWorkspaceResponsePercentileCurveItem[];
  portfolio?: RiskWorkspaceResponsePortfolio;
  projection_version: number;
  regimes?: RiskWorkspaceResponseRegimes;
  risks?: RiskWorkspaceResponseRisksItem[];
  scenario?: RiskWorkspaceResponseScenario;
  source_snapshot_ids?: string[];
  tail?: RiskWorkspaceResponseTail;
}